Refactored proxy code
