+++ download.js
@@ -10,8 +10,7 @@ var noop = Object.assign({
 var zlib = require('zlib')
 var util = require('./util')
 var error = require('./error')
-var url = require('url')
-var tunnel = require('tunnel-agent')
+var proxy = require('./proxy')
 var mkdirp = require('mkdirp')
 
 function downloadPrebuild (downloadUrl, opts, cb) {
@@ -46,8 +45,7 @@ function downloadPrebuild (downloadUrl, opts, cb) {
         }
 
         log.http('request', 'GET ' + downloadUrl)
-        var reqOpts = { url: downloadUrl }
-        var proxy = opts['https-proxy'] || opts.proxy
+        var reqOpts = proxy({ url: downloadUrl }, opts)
 
         if (opts.token) {
           reqOpts.url += '?access_token=' + opts.token
@@ -55,26 +53,6 @@ function downloadPrebuild (downloadUrl, opts, cb) {
             'User-Agent': 'simple-get',
             'Accept': 'application/octet-stream'
           }
-        }
-
-        if (proxy) {
-          var parsedDownloadUrl = url.parse(downloadUrl)
-          var parsedProxy = url.parse(proxy)
-          var uriProtocol = (parsedDownloadUrl.protocol === 'https:' ? 'https' : 'http')
-          var proxyProtocol = (parsedProxy.protocol === 'https:' ? 'Https' : 'Http')
-          var tunnelFnName = [uriProtocol, proxyProtocol].join('Over')
-          reqOpts.agent = tunnel[tunnelFnName]({
-            proxy: {
-              host: parsedProxy.hostname,
-              port: +parsedProxy.port,
-              proxyAuth: parsedProxy.auth
-            }
-          })
-          log.http('request', 'Proxy setup detected (Host: ' +
-            parsedProxy.hostname + ', Port: ' +
-            parsedProxy.port + ', Authentication: ' +
-            (parsedProxy.auth ? 'Yes' : 'No') + ')' +
-            ' Tunneling with ' + tunnelFnName)
         }
 
         var req = get(reqOpts, function (err, res) {
